% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumer-package.R
\docType{package}
\name{resumer-package}
\alias{resumer}
\alias{resumer-package}
\title{resumer: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
