% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumer-package.R
\docType{package}
\name{resumr-package}
\alias{resumr}
\alias{resumr-package}
\title{resumr: A neat resume (CV) template}
\description{
A nice formatted template for resumes in Rmarkdown
}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Michael DeCorte \email{mrd@sun.soe.clarkson.edu} [copyright holder]
}

}
\keyword{internal}
